(measureAcc) manually merging the modified from 'develop_bgl5'.
