+++ measureAcc/NomenCleaner/HLA_ALLELE_TABLE.imgt3320.hat
