+++ measureAcc/data/SameExon234.MERGED.prot.txt
