+++ measureAcc/measureAccuracy.py
@@ -1,10 +1,112 @@
+#-*- coding: utf-8 -*-
 import os, sys, re
 import pandas as pd
+
+from os.path import exists, join, dirname, basename
+
+from measureAcc.NomenCleaner.NomenCleaner import HATK_NomenCleaner
+from measureAcc.src.SieveCHPED import SieveCHPED
+from measureAcc.src.ALLELES2HPED import ALLELES2HPED
+from measureAcc.src.HPED_DRB1_1454to1401 import HPED_DRB1_1454to1401
+
+from src.CookHLAError import CookHLAInputPreparationError
 
 HLA_names = ["A", "B", "C", "DPA1", "DPB1", "DQA1", "DQB1", "DRB1"]
 Meta_Info = ['FID', 'IID', 'PID', 'MID', 'Sex', 'Phe']
 
 p_HLA_allele = re.compile(r'\w+\*\d{2,3}(:\d{2,3})*')
+
+
+
+
+class CookHLA_measureAcc(object):
+
+    def __init__(self, _answer, _imputed, _out, _allele_group='measureAcc/data/SameExon234.MERGED.nuc.txt'):
+
+        """
+        1. imputed *.alleles
+        2. answer *.alleles
+
+        (1) ALLELES2HPED.py
+
+        (2) DRB1 1454 -> 1401
+
+        (3) NomenCleaner
+
+        (4) SieveCHEPD.py
+
+        (5) measureAccuracy_v3.5
+
+        """
+
+        ## Exception Handling here.
+
+        # (1) answer file
+        if not exists(_answer):
+            raise CookHLAInputPreparationError("Given answer file can't be found.('{}')".format(_answer))
+
+        if not (_answer.endswith('.alleles') or _answer.endswith('.hped') or _answer.endswith('.Marked.chped')):
+            # No *.chped alone
+            raise CookHLAInputPreparationError("Given answer file must have file extension "
+                                               "either '*.alleles', '*.hped' or '*.Marked.chped'.")
+
+        # (2) imputed file
+        if not exists(_imputed):
+            raise CookHLAInputPreparationError("Given imputed file can't be found.('{}')".format(_imputed))
+
+        if not (_imputed.endswith('.alleles') or _imputed.endswith('.hped') or _imputed.endswith('.Marked.chped')):
+            # No *.chped alone
+            raise CookHLAInputPreparationError("Given imputed file must have file extension "
+                                               "either '*.alleles', '*.hped' or '*.Marked.chped'.")
+
+
+        # (3) _allele_group file
+        if not exists(_allele_group):
+            raise CookHLAInputPreparationError("Given Allele Group file can't be found.('{}')".format(_allele_group))
+
+
+
+        ## measureAccuracy_v3.5
+        answer_Marked_chped = self.ConvertToMarkedCHPED(_answer, dirname(_out))
+        imputed_Marked_chped = self.ConvertToMarkedCHPED(_imputed, dirname(_out))
+
+        self.accuracy = measureAccuracy(answer_Marked_chped, imputed_Marked_chped, _out, _allele_group)
+
+
+
+
+    def ConvertToMarkedCHPED(self, _f, _out_dir):
+
+        if _f.endswith('.alleles'):
+            # ALLELES2HPED
+            t_out = join(_out_dir, re.sub(r'\.alleles$', '', basename(_f)))
+            _f = ALLELES2HPED(_f, t_out, _f_HLA_DRB1_1454to1401=True)
+            # print("alleles -> hped: {}".format(_f))
+
+        if _f.endswith('.hped'):
+
+            # (1) HLA_DRB1 1454 to 1401
+            _f = HPED_DRB1_1454to1401(_f, re.sub(r'hped$', '1454to1401.hped', _f))
+
+
+            # (2) NomenCleaner
+            t1_out = re.sub(r'hped$', 'imgt3320.4field', _f)
+            t = HATK_NomenCleaner(_f, "measureAcc/NomenCleaner/HLA_ALLELE_TABLE.imgt3320.hat", '3320', t1_out,
+                      __f_NoCaption=False, __leave_NotFound=False,
+                      __oneF=False, __twoF=False, __threeF=False, __fourF=True, __Ggroup=False, __Pgroup=False)
+            # print("hped -> chped: {}".format(t.chped))
+
+            # (3) SieveCHPED
+            _f = SieveCHPED(_f, t.chped, t1_out)
+            # print("chped -> Marked.chped: {}".format(_f))
+
+        if _f.endswith('.Marked.chped'):
+            return _f
+        else:
+            return '-1'
+
+
+
 
 def whichGroup(_allele, _allele_gruop):
