+++ measureAcc/src/SieveCHPED.py
@@ -1,3 +1,4 @@
+#-*- coding: utf-8 -*-
 import os, sys, re
 import pandas as pd
